SICM6Acodes/1. predictor/requirements.txt
python 3.7;
pytorch 1.0;
numpy;
wx;
sklearn;
pylab;
os;


